int salary changed to double

+++ dependencyInjection/src/main/java/com/example/dependencyInjection/model/JuniorDeveloper.java
@@ -3,7 +3,7 @@ package com.example.dependencyInjection.model;
 public class JuniorDeveloper extends Developer{
 
 
-    public JuniorDeveloper(int id, String name, int salary) {
+    public JuniorDeveloper(int id, String name, double salary) {
         super(id, name, salary, Experience.JUNIOR);
     }
 }
